fix: Use theme variables and rename overline component file
