+++ src/routes/safe/components/Balances/SendModal/screens/ContractInteraction/SendCustomTx/index.tsx
@@ -26,6 +26,7 @@ import { getNativeCurrency } from 'src/config'
 import { EthAddressInput } from '../EthAddressInput'
 import { ensResolver, formMutators } from '../utils'
 import Buttons from '../Buttons'
+import { getStepTitle } from 'src/routes/safe/components/Balances/SendModal/utils'
 
 export interface CreatedTx {
   contractAddress: string
@@ -69,7 +70,7 @@ const SendCustomTx = ({ initialValues, isABI, onClose, onNext, switchMethod }: P
         <Paragraph className={classes.manage} noMargin weight="bolder">
           Contract interaction
         </Paragraph>
-        <Paragraph className={classes.annotation}>Step 1 of 2</Paragraph>
+        <Paragraph className={classes.annotation}>{getStepTitle(1, 2)}</Paragraph>
         <IconButton disableRipple onClick={onClose}>
           <Close className={classes.closeIcon} />
         </IconButton>
